+++ packages/sphinx-test-reports/sphinxcontrib/test_reports/directives/test_suite.py
@@ -104,13 +104,13 @@ class TestSuiteDirective(TestCommonDirective):
             tags=self.test_tags,
             status=self.test_status,
             collapse=self.collapse,
-            file=self.test_file_given,
             suite=suite["name"],
             cases=cases,
             passed=passed,
             skipped=skipped,
             failed=failed,
             errors=errors,
+            **{self.report_file_field(): self.test_file_given},
             **self.extra_options,
         )